python_modules/qjazz-server/requirements.txt
PyYAML==6.0.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
annotated-types==0.7.0
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.7
dnspython==2.7.0
frozenlist==1.5.0
grpcio-health-checking==1.68.1
grpcio-tools==1.68.1
grpcio==1.68.1
idna==3.10
jsondiff==2.2.1
minio==7.2.12
multidict==6.1.0
propcache==0.2.1
protobuf==5.29.1
pycparser==2.22
pycryptodome==3.21.0
pydantic-settings==2.7.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
ruamel.yaml.clib==0.2.12
ruamel.yaml==0.18.6
setuptools==75.6.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
